--- gatsby-browser.js
@@ -1,12 +1,0 @@
-/**
- * Implement Gatsby's Browser APIs in this file.
- *
- * See: https://www.gatsbyjs.org/docs/browser-apis/
- */
-
-// You can delete this file if you're not using it
-import 'normalize.css/normalize.css';
-import 'rc-footer/assets/index.less';
-import 'antd/dist/antd.css';
-import 'prism-themes/themes/prism-base16-ateliersulphurpool.light.css';
-import 'prismjs/plugins/command-line/prism-command-line.css';
