+++ gatsby-config.js
@@ -1,4 +1,10 @@
 module.exports = {
+  plugins: [
+    {
+      resolve: `@antv/gatsby-theme-antv`,
+      options: {},
+    },
+  ],
   siteMetadata: {
     title: `AntV`,
     siteUrl: `https://antvis.github.io`,
@@ -15,6 +21,7 @@ module.exports = {
           en: 'Specification',
         },
         order: 0,
+        redirect: 'getting-started',
       },
       {
         slug: 'specification/category',
@@ -31,6 +38,7 @@ module.exports = {
           en: 'other',
         },
         order: 0,
+        redirect: 'getting-started',
       },
       {
         slug: 'other/category',
@@ -42,90 +50,4 @@ module.exports = {
       },
     ],
   },
-  plugins: [
-    `gatsby-plugin-react-helmet`,
-    {
-      resolve: `gatsby-source-filesystem`,
-      options: {
-        name: `docs`,
-        path: `${__dirname}/docs`,
-      },
-    },
-    {
-      resolve: 'gatsby-plugin-i18n',
-      options: {
-        langKeyDefault: 'zh',
-        useLangKeyLayout: false,
-      },
-    },
-    {
-      resolve: `gatsby-transformer-remark`,
-      options: {
-        plugins: [
-          {
-            resolve: `gatsby-remark-prettier`,
-            options: {
-              // Look for local .prettierrc file.
-              // The same as `prettier.resolveConfig(process.cwd())`
-              usePrettierrc: true,
-              // Overwrite prettier options, check out https://prettier.io/docs/en/options.html
-              prettierOptions: {},
-            },
-          },
-          {
-            resolve: `gatsby-remark-prismjs`,
-            options: {
-              inlineCodeMarker: '±',
-            },
-          },
-          {
-            resolve: 'gatsby-remark-external-links',
-            options: {
-              target: '_self',
-              rel: 'nofollow',
-            },
-          },
-          `gatsby-remark-autolink-headers`,
-          `gatsby-remark-reading-time`,
-        ],
-      },
-    },
-    `gatsby-transformer-sharp`,
-    `gatsby-plugin-sharp`,
-    `gatsby-plugin-offline`,
-    {
-      resolve: `gatsby-plugin-manifest`,
-      options: {
-        name: `gatsby-starter-default`,
-        short_name: `starter`,
-        start_url: `/`,
-        background_color: `#663399`,
-        theme_color: `#663399`,
-        display: `minimal-ui`,
-        icon: `src/images/favicon.png`, // This path is relative to the root of the site.
-      },
-    },
-    `gatsby-plugin-typescript`,
-    {
-      resolve: 'gatsby-plugin-less',
-      options: {
-        javascriptEnabled: true,
-      },
-    },
-    {
-      resolve: 'gatsby-plugin-antd',
-      options: {
-        style: true,
-      },
-    },
-    {
-      resolve: `gatsby-plugin-google-analytics`,
-      options: {
-        trackingId: `UA-148148901-1`,
-      },
-    },
-    `gatsby-plugin-catch-links`,
-    `gatsby-plugin-nprogress`,
-    `gatsby-plugin-remove-trailing-slashes`,
-  ],
 };
