--- jest-preprocess.js
@@ -1,4 +1,0 @@
-const babelOptions = {
-  presets: ['babel-preset-gatsby'],
-};
-module.exports = require('babel-jest').createTransformer(babelOptions);
