--- jest.setup.js
@@ -1,1 +1,0 @@
-require(`@testing-library/jest-dom/extend-expect`);
