--- loadershim.js
@@ -1,3 +1,0 @@
-global.___loader = {
-  enqueue: jest.fn(),
-};
