+++ package.json
@@ -1,5 +1,6 @@
 {
-  "name": "@antv/gatsby-theme",
+  "name": "antv.alipay.com",
+  "private": true,
   "description": "Ant Visualization solution home page",
   "version": "0.1.0",
   "author": "AntV Team <antv@antfin.com>",
@@ -17,72 +18,25 @@
     "url": "https://github.com/antvis/antvis.github.io/issues"
   },
   "dependencies": {
-    "@hot-loader/react-dom": "^16.9.0+4.12.11",
-    "@testing-library/jest-dom": "^4.1.0",
-    "@testing-library/react": "^9.1.4",
-    "@types/classnames": "^2.2.9",
-    "@types/jest": "^24.0.18",
-    "@types/lodash-es": "^4.17.3",
-    "@types/react": "^16.9.2",
-    "@types/react-dom": "^16.9.0",
-    "@types/react-helmet": "^5.0.10",
-    "antd": "^3.23.3",
-    "babel-jest": "^24.9.0",
-    "babel-preset-gatsby": "^0.2.13",
-    "classnames": "^2.2.6",
-    "gatsby": "^2.15.16",
-    "gatsby-image": "^2.2.19",
-    "gatsby-plugin-antd": "^2.0.2",
-    "gatsby-plugin-catch-links": "^2.1.12",
-    "gatsby-plugin-google-analytics": "^2.1.16",
-    "gatsby-plugin-i18n": "^1.0.1",
-    "gatsby-plugin-less": "^3.0.6",
-    "gatsby-plugin-manifest": "^2.2.16",
-    "gatsby-plugin-nprogress": "^2.1.9",
-    "gatsby-plugin-offline": "^3.0.7",
-    "gatsby-plugin-react-helmet": "^3.1.7",
-    "gatsby-plugin-remove-trailing-slashes": "^2.1.9",
-    "gatsby-plugin-sharp": "^2.2.22",
-    "gatsby-plugin-typescript": "^2.1.8",
-    "gatsby-remark-autolink-headers": "^2.1.13",
-    "gatsby-remark-external-links": "^0.0.4",
-    "gatsby-remark-prettier": "^1.0.0",
-    "gatsby-remark-prismjs": "^3.3.16",
-    "gatsby-remark-reading-time": "^1.0.1",
-    "gatsby-source-filesystem": "^2.1.22",
-    "gatsby-transformer-remark": "^2.6.24",
-    "gatsby-transformer-sharp": "^2.2.14",
-    "husky": "^3.0.5",
-    "i18next": "^17.0.16",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^24.9.0",
-    "less": "^3.10.3",
-    "lodash-es": "^4.17.15",
-    "moment": "^2.24.0",
-    "normalize.css": "^8.0.1",
+    "@antv/gatsby-theme-antv": "^0.1.1",
+    "@types/react": "^16.9.5",
+    "@types/react-dom": "^16.9.1",
+    "antd": "^3.23.6",
+    "gatsby": "^2.15.36",
+    "husky": "^3.0.8",
     "prettier": "^1.18.2",
     "pretty-quick": "^1.11.0",
-    "prism-themes": "^1.2.0",
-    "prismjs": "^1.17.1",
-    "ptz-i18n": "^1.0.0",
-    "rc-footer": "^0.5.0",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
-    "react-github-corner": "^2.3.0",
-    "react-helmet": "^5.2.1",
-    "react-i18next": "^10.13.1",
-    "react-test-renderer": "^16.9.0",
-    "ts-jest": "^24.1.0",
-    "typescript": "^3.6.3"
+    "react": "^16.10.2",
+    "react-dom": "^16.10.2",
+    "typescript": "^3.6.4"
   },
   "license": "MIT",
   "scripts": {
+    "start": "npm run develop",
     "build": "gatsby build",
     "develop": "gatsby develop --open",
     "format": "prettier --write \"**/*.{ts,tsx,js,jsx,json,md,less}\"",
-    "start": "npm run develop",
-    "serve": "gatsby serve",
-    "test": "jest"
+    "serve": "gatsby serve"
   },
   "husky": {
     "hooks": {
