+++ site/images/gatsby-astronaut.png
