+++ site/pages/index.en.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Skeleton } from 'antd';
-import Layout from '../components/layout';
-import SEO from '../components/seo';
+import Layout from '@antv/gatsby-theme-antv/site/components/layout';
+import SEO from '@antv/gatsby-theme-antv/site/components/seo';
 
 const IndexPage = () => (
   <Layout>
-    <SEO title="蚂蚁数据可视化" lang="zh" />
+    <SEO title="Home" lang="en" />
     <div style={{ margin: '0 auto', padding: '0 80px' }}>
-      中文首页
+      Home Page
       <Skeleton />
       <Skeleton />
       <Skeleton />
