--- src/components/article.module.less
@@ -1,9 +1,0 @@
-.article {
-  padding: 0 16px 0 48px;
-  min-height: 600px;
-
-  h2 {
-    font-size: 36px;
-    margin-top: 0;
-  }
-}
