--- src/components/header.module.less
@@ -1,67 +1,0 @@
-.header {
-  height: 64px;
-  line-height: 64px;
-  background: #fff;
-  box-shadow: 0 2px 8px rgba(229, 229, 229, 0.5);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 60px;
-  position: relative;
-  z-index: 1;
-
-  .left {
-    display: flex;
-    align-items: center;
-
-    h1 {
-      margin: 0;
-
-      img {
-        display: block;
-      }
-    }
-
-    .divider {
-      width: 1px;
-      height: 24px;
-      background-color: #ccc;
-      display: inline-block;
-      margin: 0 24px;
-    }
-  }
-
-  .nav {
-    display: flex;
-    align-items: center;
-
-    .menu {
-      font-weight: 500;
-      color: rgba(0, 0, 0, 0.65);
-      margin: 0;
-      padding: 0;
-      list-style: none;
-
-      li {
-        display: inline-block;
-        margin-right: 32px;
-        line-height: 60px;
-
-        a {
-          color: rgba(0, 0, 0, 0.65);
-          transition: all 0.3s;
-          display: block;
-          border-top: 2px solid transparent;
-
-          &:hover {
-            color: #1890ff;
-          }
-        }
-      }
-
-      .active {
-        border-top: 2px solid #1890ff;
-      }
-    }
-  }
-}
