--- src/components/layout.module.less
@@ -1,15 +1,0 @@
-:global {
-  .gatsby-highlight-code-line {
-    background-color: #feb;
-    display: block;
-    margin-right: -1em;
-    margin-left: -1em;
-    padding-right: 1em;
-    padding-left: 0.75em;
-    border-left: 0.25em solid #f99;
-  }
-}
-
-.main {
-  margin: 24px 0;
-}
