--- src/components/search.module.less
@@ -1,5 +1,0 @@
-.input {
-  border: none;
-  height: 32px;
-  outline: none;
-}
