--- src/externals.d.ts
@@ -1,3 +1,0 @@
-declare module '*.less';
-
-declare module 'ptz-i18n';
