--- src/templates/markdown.module.less
@@ -1,165 +1,0 @@
-@import '~antd/es/style/themes/default.less';
-
-.markdown {
-  font-size: 14px;
-  line-height: 2;
-  color: #222;
-
-  > img {
-    max-width: ~'calc(100% - 32px)';
-    margin: 34px 0;
-  }
-
-  h1 {
-    margin-top: 8px;
-    margin-bottom: 20px;
-    font-weight: 500;
-    font-size: 30px;
-    font-family: Avenir, @font-family, sans-serif;
-    line-height: 38px;
-  }
-
-  h2 {
-    font-size: 24px;
-    line-height: 32px;
-  }
-
-  h2,
-  h3,
-  h4,
-  h5,
-  h6 {
-    clear: both;
-    margin: 1.6em 0 0.6em;
-    font-weight: 500;
-    font-family: Avenir, @font-family, sans-serif;
-  }
-
-  h3 {
-    font-size: 18px;
-  }
-  h4 {
-    font-size: 16px;
-  }
-  h5 {
-    font-size: 14px;
-  }
-  h6 {
-    font-size: 12px;
-  }
-
-  hr {
-    clear: both;
-    height: 1px;
-    margin: 56px 0;
-    background: @border-color-split;
-    border: 0;
-  }
-
-  p,
-  pre {
-    margin: 1em 0;
-  }
-
-  ul,
-  ol {
-    margin: 0;
-    padding: 0;
-  }
-
-  ul > li {
-    margin-left: 20px;
-    padding-left: 4px;
-    list-style-type: circle;
-    &:empty {
-      display: none;
-    }
-  }
-
-  ol > li {
-    margin-left: 20px;
-    padding-left: 4px;
-    list-style-type: decimal;
-  }
-
-  ul > li > p,
-  ol > li > p {
-    margin: 0.2em 0;
-  }
-
-  strong,
-  b {
-    font-weight: 500;
-  }
-
-  > table {
-    width: 100%;
-    margin: 8px 0 16px;
-    empty-cells: show;
-    border: 1px solid @border-color-split;
-    border-collapse: collapse;
-    border-spacing: 0;
-  }
-
-  > table th {
-    color: #5c6b77;
-    font-weight: 500;
-    white-space: nowrap;
-    background: rgba(0, 0, 0, 0.02);
-  }
-
-  > table th,
-  > table td {
-    padding: 16px 24px;
-    text-align: left;
-    border: 1px solid @border-color-split;
-  }
-
-  blockquote {
-    margin: 1em 0;
-    padding-left: 0.8em;
-    color: @text-color-secondary;
-    font-size: 90%;
-    border-left: 4px solid @border-color-split;
-  }
-
-  blockquote p {
-    margin: 0;
-  }
-
-  > br,
-  > p > br {
-    clear: both;
-  }
-}
-
-.editOnGtiHubButton {
-  margin-left: 0.5em;
-  font-size: 16px;
-  color: #999;
-  vertical-align: middle;
-  position: relative;
-  top: -2px;
-}
-
-.main {
-  width: calc(100% - 160px);
-  padding-right: 24px;
-  padding-bottom: 60px;
-}
-
-.toc {
-  width: 160px;
-  border-left: 2px solid #eee;
-  float: right;
-  font-size: 13px;
-
-  ul > li {
-    list-style: none;
-    padding: 0;
-    margin-left: 16px;
-    a {
-      color: rgba(0, 0, 0, 0.65);
-    }
-  }
-}
